Move restoration_db_handler to test_helpers_internal
