+++ ethcore/src/test_helpers.rs
@@ -35,11 +35,8 @@ use spec::Spec;
 use state_db::StateDB;
 use state::*;
 use std::sync::Arc;
-use std::path::Path;
 use transaction::{Action, Transaction, SignedTransaction};
 use views::BlockView;
-use kvdb::{KeyValueDB, KeyValueDBHandler};
-use kvdb_rocksdb::{Database, DatabaseConfig};
 
 /// Creates test block with corresponding header
 pub fn create_test_block(header: &Header) -> Bytes {
@@ -401,21 +398,4 @@ impl ChainNotify for TestNotify {
 		};
 		self.messages.write().push(data);
 	}
-}
-
-/// Creates new instance of KeyValueDBHandler
-pub fn restoration_db_handler(config: DatabaseConfig) -> Box<KeyValueDBHandler> {
-	use kvdb::Error;
-
-	struct RestorationDBHandler {
-		config: DatabaseConfig,
-	}
-
-	impl KeyValueDBHandler for RestorationDBHandler {
-		fn open(&self, db_path: &Path) -> Result<Arc<KeyValueDB>, Error> {
-			Ok(Arc::new(Database::open(&self.config, &db_path.to_string_lossy())?))
-		}
-	}
-
-	Box::new(RestorationDBHandler { config })
 }
